requirements.txt
Flask
gunicorn
peewee
psycopg2-binary